Got the spike working! :-)
Had to use IValidator rather than IValidator<IAmValidatable> which means
using an obsolete method on FluentNHibernate.  Just going with that for
now...

We need to use the old CompositeValidator for all validators now (even
single ones) as it is responsible for converting the validator to
IValidator<IAmValidatable>.

Needs cleaned up and specs put around it.