Implementing an observable collection - look at description
In WindowsBase you'll find an implementation of this - but I don't want
to bring in a dependency to something that is first of all meant for
frontend and secondly is a Windows only assembly basically. Besides,
even though the name is the same now - it might change.

On a design note; instead of having just what the
INotifyCollectionChanged interface is giving - I've added specific
events and delegates for the 3 events that can occur. This is in my
opinion much clearer than INCC.