Refactoring TypeDiscoverer a bit to accommodate new needs + changing signatures
Basically we are going to need to be able to find types from any enumerable of types. TypeDiscoverer had a bunch of this functionality we wanted but coupled together with how it wants to do things. Extracting out the finding of types part into its own ITypeFinder enables us to reuse the finding bits.

Worth mentioning is that all signatures returning arrays of Type are now returning IEnumerable<Type> instead. This affected some implementations!