Update quickstart.md
Not necessary with with the special package version. The referenced packages (Ninject and FluentValidation) are not on NuGet any more, and a clean install of the QuickStart without installing Ninject and FluentValidation first was the only that worked!

I tried installing Ninject and FluentValidation packages first but that only resulted in compilation errors and server errors during execution.